Deprecate old file.
